Realized I should have committed the CLI. Oops.
The CLI uses picocli (might be overkill but it seems to be handling a lot of the obnoxious arg-parsing stuff for me) and correctly launches the actual script, plus hooks to a database correctly. Might want to add an option to disable the 'ask for confirmation' bit.

File path stuff seems more resilient than I expected- it doesn't explode if you terminate the argument in a / and it works with both forward and backslashes- but I doubt it's portable to non-Windows.