Seems to build a correct HashMap of {token, name} pairs for each album or face. Doesn't actually tag them yet. I'll probably look for the 'albums' line for a file and add a tag to tags[] if it's found, then remove the isNotJpg check. (replace with a metaLine != null?)
Also, this process is... kinda slow on my folder with 16k files in it, before I even had it start tagging jpgs in albums. Not wild about it. Though I'm not sure how I'd meaningfully speed up the tagging process and realistically, this is something you're only going to run once, so optimization isn't necessarily critical.